refactor: ♻️ (api) interface-adapter
